chore: codebase folder tree refactoring
